requirements.txt
psycopg2
pillow
django-celery
kombu
pytz
pycurl
django-admin-tools
South
django-mailer
django-registration
django_extensions
django_nose
django_mobile
django-crispy-forms
django-generic-m2m
django-sendsms
django-sekizai
django-form-designer
django-filer
django-cms
django-reversion
django-tinymce
cmsplugin-filer
